+++ shared_code/__init__.py
@@ -1,3 +1,3 @@
 """Initialize package"""
 
-from shared_code import get_config, cosmosdb_module, aio_helper, date_time_helper, schemas
+from shared_code import get_config, cosmosdb_module, aio_helper, date_time_helper, schemas, utils
